#46 confirm dialog / example confirm

+++ frontend/src/app/mymeetups/own-meetups/own-meetups.component.ts
@@ -2,6 +2,8 @@ import {Component, OnInit} from '@angular/core';
 import {BusinessService} from '../../core/business.service';
 import {Meetup, Hall} from '@chumm-uffa/interface';
 import {Util} from '../../shared/util';
+import {ConfirmDialogComponent} from '../../material/confirm-dialog/confirm-dialog.component';
+import {MatDialog} from '@angular/material';
 
 @Component({
   selector: 'app-own-meetups',
@@ -12,7 +14,8 @@ export class OwnMeetupsComponent implements OnInit {
   meetups: Meetup[] = [];
   halls: Hall[] = [];
 
-  constructor(private businessService: BusinessService) {
+  constructor(private businessService: BusinessService,
+              private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -30,6 +33,15 @@ export class OwnMeetupsComponent implements OnInit {
 
   deleteMeetup(event, meetupId: string): void {
     event.stopPropagation();
-    this.businessService.deleteMeetup(meetupId).subscribe(_ => this.getMeetups());
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent,
+      {data: {confirmText: 'ownMeetups.dialog.signOff_text', confirmTitle: 'ownMeetups.dialog.signOff_title'}});
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 'yes') {
+        this.businessService.deleteMeetup(meetupId).subscribe(_ => this.getMeetups());
+      }
+    });
+
   }
 }
